Add concept of "sticky inputs"
By right clicking an input port, you can toggle its "stickiness". Sticky
input ports are not cleared after the patch has been executed. With
this, you can fill a port once, and have the last value always
available. This is useful, for example, when adding a constant to every
value in an array.

Sticky values should only trigger execution when the value changes.